Deprecate RequestAndAcceptMatcher to AcceptMatcher
Functionality will be restored via usage of AndMatcher which is to added
by the next commit.
